location request at the start of app
