+++ Mood+/app/src/main/java/com/mood/jenaPlus/MarkerActivity.java
@@ -10,6 +10,7 @@ import android.os.Bundle;
 import android.util.Log;
 import android.widget.Toast;
 
+import com.google.android.gms.location.places.ui.PlacePicker;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
@@ -25,14 +26,10 @@ public class MarkerActivity extends FragmentActivity implements
         GoogleMap.OnMarkerClickListener,
         OnMapReadyCallback {
 
-    private Marker testMarker;
-
-    private GoogleMap mMap;
-
-    private Marker followingMarker;
     ArrayList<Mood> moodListLocation;
     LatLng allLatLng;
     ArrayList<Mood> userMoodLocation;
+    int PLACE_PICKER_REQUEST = 1;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -43,20 +40,23 @@ public class MarkerActivity extends FragmentActivity implements
                 (SupportMapFragment) getSupportFragmentManager().findFragmentById(R.id.map);
         mapFragment.getMapAsync(this);
 
+
+        PlacePicker.IntentBuilder builder = new PlacePicker.IntentBuilder();
+
+        //startActivityForResult(builder.build(this), PLACE_PICKER_REQUEST);
+
     }
 
     /** Called when the map is ready. */
     @Override
     public void onMapReady(GoogleMap map) {
-        mMap = map;
-
-        //moodListLocation = (ArrayList<Mood>) getIntent().getSerializableExtra("participant_moodProvider");
-        //userMoodLocation = (ArrayList<Mood>) getIntent().getSerializableExtra("user_moodProvider");
+        GoogleMap mMap = map;
+        Marker followingMarker;
 
         if( getIntent().hasExtra("user_moodProvider")) {
             userMoodLocation = (ArrayList<Mood>) getIntent().getSerializableExtra("user_moodProvider");
-            Log.i("TAGGGGGG","PASSING USER MOOD LOCATION" + userMoodLocation);
-            Log.i("SECOND TAG","EMPTY"+moodListLocation);
+            //Log.i("TAGGGGGG","PASSING USER MOOD LOCATION" + userMoodLocation);
+            //Log.i("SECOND TAG","EMPTY"+moodListLocation);
 
             for (Mood mood: userMoodLocation){
                 allLatLng = new LatLng(mood.getLatitude(), mood.getLongitude());
@@ -73,12 +73,12 @@ public class MarkerActivity extends FragmentActivity implements
         }
         else if(getIntent().hasExtra("participant_moodProvider")){
             moodListLocation = (ArrayList<Mood>) getIntent().getSerializableExtra("participant_moodProvider");
-            Log.i("TAGGGGGG","PASSING Participant moods LOCATION" + moodListLocation);
-            Log.i("SECOND TAG","EMPTY"+userMoodLocation);
+            //Log.i("TAGGGGGG","PASSING Participant moods LOCATION" + moodListLocation);
+            //Log.i("SECOND TAG","EMPTY"+userMoodLocation);
 
             for (Mood mood: moodListLocation){
                 //Log.i("LATLNG!!!!!!","Contents of arrayLocation: " + mood.getLatitude()+mood.getLongitude() );
-                Log.i("MOODS ID!!!!!!",mood.getId() + mood.getUserName());
+                //Log.i("MOODS ID!!!!!!",mood.getId() + mood.getUserName());
                 allLatLng = new LatLng(mood.getLatitude(), mood.getLongitude());
 
                 // Creating markers
